Added callbacks infrastructure
This commit adds the possibility to implement callbacks. This low-level
interface allows to interact with the command line utility of the
MainWindow. Two callbacks have been added by this: The `script`
callback, to run a script in the console, and the `command` callback,
to run a python command in the console.